backend/requirements.txt
boto3==1.34.34
bcrypt==4.1.2
PyJWT==2.8.0
python-dotenv==1.0.0
